Add files via upload
All that jazz with the gui which is needed to display the jsons output. 
I also changed the colors to prevent my eyes from burning out while looking at it.
The first screen got crowded so I moved the filtering over to another tab.
It worked the last time I ran it, so hopefully it works again with other people.